% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_indicator}
\alias{add_indicator}
\title{Add an indicator to your strategy}
\usage{
add_indicator(strategy_object, indicator_name, generator, generator_args)
}
\arguments{
\item{strategy_object}{a \code{\link{strategy}} object}

\item{indicator_name}{the desired column name of the indicator, used to build signals}

\item{generator}{the function used to generate the indicator value}

\item{generator_args}{a named list of parameters to use with the generator function}
}
\value{
a strategy object with indicator applied to all securities
}
\description{
Calculates an indicator, and adds it to a column in each security's timeseries data.
}
