% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_signal}
\alias{add_signal}
\title{Add a signal to your strategy}
\usage{
add_signal(strategy_object, signal_name, signal, direction = "buy",
  crossover = TRUE)
}
\arguments{
\item{strategy_object}{a \code{\link{strategy}} object}

\item{signal_name}{the desired column name of the signal, used to build rules}

\item{signal}{a logical argument based on the indicator columns of the security object}

\item{direction}{the direction of the market bet, buy or sell}

\item{crossover}{defines whether signal should be a crossover signal. If TRUE, the signal will only occur on first bar of each run of signal condition being met.}
}
\value{
a strategy object with signal applied to all securities
}
\description{
Determines whether a signal has occurred based on one or more columns in the strategy data, and adds it to a column in each security's timeseries data.
}
