% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{chart_equity_curve}
\alias{chart_equity_curve}
\title{Chart a strategy's equity curve}
\usage{
chart_equity_curve(strategy_object)
}
\arguments{
\item{strategy_object}{a \code{\link{strategy}} object}
}
\description{
From a backtested strategy object, generate a plot of the equity curve
}
