% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{chart_optimizer}
\alias{chart_optimizer}
\title{Chart the results of a strategy optimization}
\usage{
chart_optimizer(strategy_object)
}
\arguments{
\item{strategy_object}{a strategy object that's been produced by the optimize_strategy() function}
}
\value{
a ggplot2 object
}
\description{
Generates a chart of different tested parameter values versus the equity those parameters produced. Maximum
equity is annotated with an orange dot.
}
