% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{chart_signals}
\alias{chart_signals}
\title{Chart trading signals}
\usage{
chart_signals(strategy_object, ticker, use_price = "CLOSE")
}
\arguments{
\item{strategy_object}{a \code{\link{strategy}} object}

\item{ticker}{the ticker symbol to chart}

\item{use_price}{price column to chart}
}
\description{
Generate a price chart with overlaid buy and sell signals for a given ticker from the strategy universe.
}
