% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_engine.R
\name{compile_strategy}
\alias{compile_strategy}
\title{Compile a strategy for backtesting}
\usage{
compile_strategy(strategy_object, signals)
}
\arguments{
\item{strategy_object}{a strategy object}

\item{signals}{a list of signal columns}
}
\value{
a strategy object
}
\description{
Determines buy and sell trades based on a strategy's signals.
}
