% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxcapacitoR-package.R
\docType{package}
\name{fluxcapacitoR-package}
\alias{fluxcapacitoR}
\alias{fluxcapacitoR-package}
\title{fluxcapacitoR: Flux-Focused Age Modeling for Layered Sediments}
\description{
An age modeling strategy focused on flux and accumulation rates in layered sediments. Builds varve-informed age-ensemble priors that combine low-frequency monotone spline fits to dated horizons with high-frequency simulated annual-layer (varve) variability, evaluates age ensembles against calibrated age distributions, and simulates synthetic varve thickness series with AR(1) or long-memory persistence. Part of the 'geoChronR' package family.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickmckay/fluxcapacitoR}
  \item Report bugs at \url{https://github.com/nickmckay/fluxcapacitoR/issues}
}

}
\author{
\strong{Maintainer}: Nicholas McKay \email{nick@nau.edu} (\href{https://orcid.org/0000-0003-3598-5113}{ORCID})

}
\keyword{internal}
