% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxcapacitor.R
\docType{package}
\name{fluxcapacitor}
\alias{fluxcapacitor}
\alias{fluxcapacitor-package}
\title{fluxcapacitor: A tidyverse-friendly quantitative backtesting engine.}
\description{
`fluxcapacitor` is a quantitative backtesting engine that plays nicely within the `tidyverse`.
The package gets its name from the Back to the Future series of films -- like the flux capacitor in
Doc Brown's DeLorean, this package aims to facilitate "time travel", allowing financial researchers to
simulate the results of trading strategies.
}
\details{
To learn more about fluxcapacitor, start with the vignette:
 `vignette("Report", package = "fluxcapacitor")`
}
