% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-varves.R
\name{gammify}
\alias{gammify}
\title{Transform data to a gamma distribution}
\usage{
gammify(X, shape = 1.5, mean = 1, jitter = FALSE)
}
\arguments{
\item{X}{A vector or matrix of data; each column is transformed
independently.}

\item{shape}{Shape parameter of the target gamma distribution.}

\item{mean}{Mean of the target gamma distribution.}

\item{jitter}{Add tiny random noise to break ties?}
}
\value{
A matrix the same size as \code{X}, gamma-distributed by column.
}
\description{
Transforms each column of a data matrix to a gamma distribution using the
inverse Rosenblatt transform, preserving rank order.
}
\details{
Inspired by gaussianize.R, and split.m in normal.m by Van Albada, S.J.,
Robinson P.A. (2006), Transformation of arbitrary distributions to the
normal distribution with application to EEG test-retest reliability
(J Neurosci Meth, doi:10.1016/j.jneumeth.2006.11.004). Modified from
matlab code written 26/06/2015 by Julien Emile-Geay (USC); translated to R
and jitter option added 7/06/2017 by Nick McKay (NAU).
}
