% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{information_ratio}
\alias{information_ratio}
\title{Calculate a strategy's information ratio}
\usage{
information_ratio(strategy_object, benchmark_returns)
}
\arguments{
\item{strategy_object}{a strategy object}

\item{benchmark_returns}{the benchmark returns, either numeric or a vector with the same
periodicity as the strategy object's ledger.}
}
\value{
a tibble containing the strategy's information ratio
}
\description{
The information ratio is used to measure risk-adjusted return relative to a benchmark.
}
