% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_engine.R
\name{init_strategy}
\alias{init_strategy}
\title{Initialize a strategy object}
\usage{
init_strategy(universe)
}
\arguments{
\item{universe}{a character vector of the securities to be included in the strategy.}
}
\value{
a strategy object
}
\description{
The strategy() function creates a strategy object (of S3 class fc_strategy), a container that holds all of the
data, rules, and parameters for the strategy being backtested. The function takes just one argument,
universe, which must be a character vector of the securities to be included in the strategy.
}
\details{
Further, the securities named in universe must exist as objects in the global environment containing
the data for the backtest. For instance, a strategy being run on AAPL must have "AAPL" in the universe vector and
must have an object named AAPL that contains the stock price data for AAPL.
}
\examples{
data(SPY)
strat <- init_strategy("SPY")

}
