% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimize_strategy}
\alias{optimize_strategy}
\title{Optimize a strategy parameter}
\usage{
optimize_strategy(strategy_object, generator, generator_args, optimize_range,
  signal)
}
\arguments{
\item{strategy_object}{a strategy object}

\item{generator}{the function used to generate the indicator value being optimized}

\item{generator_args}{a named list of parameters to use with the generator function. Parameters to be optimized should be denoted with a question mark (?).}

\item{optimize_range}{a range of values to optimize amongst.}

\item{signal}{the signal to evaluate the reward function on.}
}
\value{
a strategy object with an Optimized entry
}
\description{
This function can be used to find the optimal value of a range of values for a strategy parameter.
}
\details{
From the list of values in the optimize range, the function will evaluate the profitability of the strategy
under each optimize_range value. The function returns the value that results in the highest ledger value at the end
of the test.
}
