% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{return_calculate}
\alias{return_calculate}
\title{Calculate strategy returns}
\usage{
return_calculate(strategy_object, method = "arithmetic")
}
\arguments{
\item{strategy_object}{a strategy object}

\item{method}{"arithmetic" or "log" returns.}
}
\value{
A tibble of dates and per-period returns
}
\description{
This function calulates the per-period returns for a backtested strategy object. This is useful for
any portfolio analytics calculations that require portfolio values to be passed as returns. This function
can calculate either arithmetic or log returns by specifying the method argument.
}
