% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sharpe_ratio}
\alias{sharpe_ratio}
\title{Calculate a strategy's Sharpe ratio}
\usage{
sharpe_ratio(strategy_object, risk_free = 0, scale = 1)
}
\arguments{
\item{strategy_object}{a strategy object}

\item{risk_free}{the risk-free rate, either numeric or a vector with the same periodicity as the strategy object's ledger.}

\item{scale}{a scaling factor.}
}
\value{
a tibble contaning the strategy's Sharpe ratio.
}
\description{
The Sharpe ratio use used to measure risk-adjusted return. This function also implements a naive form
of scaling. For instance, for daily periodicity, using 252 as the scaling factor will annualize the result.
Note that per Andrew Lo's 2003 paper
\href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=377260}{The Statistics of Sharpe Ratios}, this
method is not recommended on its own except under very special circumstances.
}
