% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{strategy_drawdowns}
\alias{strategy_drawdowns}
\title{Calulate a strategy's drawdowns}
\usage{
strategy_drawdowns(strategy_object)
}
\arguments{
\item{strategy_object}{a strategy object}
}
\value{
a tibble containing per-period drawdown statistics
}
\description{
Drawdowns measure the performance deterioration from a previous high water mark.
}
