% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{xts_to_tibble}
\alias{xts_to_tibble}
\title{Convert an xts to tidy data}
\usage{
xts_to_tibble(xts_object)
}
\arguments{
\item{xts_object}{an xts object}
}
\description{
This is an internal helper utility
}
